Modernize codebase to C11 and improve textview-chat performance
C11 / safety cleanups
---------------------

* util.{c,h}: change safe_strcpy()'s `bytes_left` parameter (and its
  internal `mbl`) from int to size_t. All callers pass sizeof(...) or
  positive constants, so behaviour is unchanged. Drop the last
  `register` keyword (nocasestrstr).

* fe-gtk3/menu.h: replace the `#if (MENU_ID_NUM < MENU_ID_FULLSCREEN)
  #error ... #endif` guard with a C11 static_assert(), and pull in
  <assert.h>.

* common/text.c: drop the `sizeof(time_t) == 4` / `sizeof(stamp) == 4`
  branches in scrollback_save/scrollback_load. Always serialise and
  parse scrollback timestamps as 64-bit so files are portable across
  platforms regardless of native time_t width.

* common/cfgfiles.c: convert ~11 unbounded strcpy() calls in
  load_default_config() to g_strlcpy(..., sizeof(field)) for
  pchat_away_reason, pchat_completion_suffix,
  pchat_gui_ulist_doubleclick, pchat_input_command_char,
  pchat_irc_logmask, pchat_irc_no_hilight, pchat_stamp_log_format,
  pchat_stamp_text_format, pchat_text_font, pchat_text_font_main and
  pchat_text_font_alternative. Convert sprintf() in cfg_get_str() and
  set_showval() to g_snprintf().

* fe-text/fe-text.c: convert two sprintf() calls writing into a
  fixed-size buffer to g_snprintf() with the remaining buffer length.

* common/util.c: convert sprintf() in errorstring() and pchat_hash()
  to g_snprintf().

* common/dcc.c: harden dcc_add_file()'s destfile construction --
  capture the malloc'd size once and use g_strlcpy/g_strlcat
  throughout, including the Windows nick-sanitising branch. Convert
  the manual `memset(&addr, 0, sizeof(addr))` in dcc_connect_sok() to
  a `struct sockaddr_in addr = { 0 };` designated initialiser.

* common/network.c, common/sysinfo/win32/backend.c, fe-gtk3/banlist.c,
  fe-gtk3/plugingui.c, common/outbound.c (cmd_send),
  common/ssl_schannel.c (pchat_ssl_backend_version): replace
  `T x; memset(&x, 0, sizeof(x));` with `T x = { 0 };` designated
  initialisers.

* common/chanopt.c: bound the strcpy() into the shared `tbuf` buffer
  with g_strlcpy().

textview-chat performance
-------------------------

src/fe-gtk3/textview-chat.c: a number of related improvements to the
GtkTextView-based chat widget:

* Coalesce auto-scroll requests. Add `scroll_idle_id` and
  `pending_scroll` to the private struct, drop the per-call ScrollData
  allocations, and route all scroll requests through
  pchat_textview_chat_request_scroll(), which schedules at most one
  idle callback per frame. The idle source is removed in finalize.

* Cache the hand and text cursors used by URL hover highlighting in
  the private struct. ensure_cursors() lazily allocates them and the
  motion/leave handlers only call gdk_window_set_cursor() when the
  cursor actually changes. Cursors are released with g_clear_object()
  in finalize.

* Drop redundant gtk_widget_queue_draw() / gtk_widget_queue_resize()
  from pchat_chat_buffer_show(); GtkTextView already invalidates
  itself when the buffer is replaced.

* Enforce `max_lines`. Add prune_buffer_to_max_lines() and call it
  after each append; previously the limit was effectively unenforced.

* Fix a latent byte/char offset bug in URL detection by using
  g_utf8_pointer_to_offset() before gtk_text_iter_forward_chars(), and
  avoid g_strndup() by temporarily NUL-terminating in place. Use
  g_ascii_isspace() instead of isspace().

* Append plain-text runs in chunks via g_string_append_len() instead
  of per-byte g_string_append_c() in
  pchat_textview_chat_append_with_formatting().

* In flush_text_with_formatting(), replace the
  gtk_text_buffer_create_mark / delete_mark dance with a cheaper
  gtk_text_iter_get_offset() + gtk_text_buffer_get_iter_at_offset().

* Rework the three timestamp-prefix wrappers
  (pchat_textview_chat_append_with_stamp,
  pchat_textview_chat_append_indent,
  pchat_chat_buffer_append_indent): use g_string_sized_new(len + 32),
  a stack-allocated `struct tm`, localtime_r/localtime_s on Windows,
  and g_string_append_len() for the formatted timestamp.

+++ src/common/util.h
@@ -73,7 +73,7 @@ void move_file (char *src_dir, char *dst_dir, char *fname, int dccpermissions);
 int token_foreach (char *str, char sep, int (*callback) (char *str, void *ud), void *ud);
 guint32 str_hash (const char *key);
 guint32 str_ihash (const unsigned char *key);
-void safe_strcpy (char *dest, const char *src, int bytes_left);
+void safe_strcpy (char *dest, const char *src, size_t bytes_left);
 void canonalize_key (char *key);
 int portable_mode (void);
 char *encode_sasl_pass_plain (char *user, char *pass);
